refactor(game): 시스템 출력을 OutputView로 분리
