+++ src/main/java/racingcar/Application.java
@@ -2,9 +2,10 @@ package racingcar;
 
 import racingcar.controller.Game;
 import racingcar.view.InputView;
+import racingcar.view.OutputView;
 
 public class Application {
     public static void main(String[] args) {
-        new Game(new InputView());
+        new Game(new InputView(), new OutputView());
     }
 }
